+++ tests/testthat/test-sql-translation.r
@@ -49,3 +49,9 @@ test_that("Subsetting always evaluated locally", {
   expect_equal(partial_eval(quote(`_var` == x[[2]])), correct)
   expect_equal(partial_eval(quote(`_var` == y[2])), correct)
 })
+
+test_that("between translated to special form (#503)", {
+
+  out <- translate_sql(between(x, 1, 2))
+  expect_equal(out, sql('"x" BETWEEN 1.0 AND 2.0'))
+})